[IGNORE] Refactor and abstract the way to interact with the database

+++ internal/api/shared/database/file/file.go
@@ -1,4 +1,17 @@
-package database
+// Copyright 2023 The Perses Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package databaseFile
 
 import (
 	"encoding/json"
@@ -10,28 +23,46 @@ import (
 	"reflect"
 	"strings"
 
-	"github.com/perses/common/etcd"
 	"github.com/perses/perses/internal/api/config"
+	databaseModel "github.com/perses/perses/internal/api/shared/database/model"
+	modelAPI "github.com/perses/perses/pkg/model/api"
+	modelV1 "github.com/perses/perses/pkg/model/api/v1"
 	"gopkg.in/yaml.v2"
 )
 
-type fileDAO struct {
-	DAO
-	folder    string
-	extension config.FileExtension
+func generateKey(kind modelV1.Kind, metadata modelAPI.Metadata) (string, error) {
+	switch m := metadata.(type) {
+	case *modelV1.ProjectMetadata:
+		return fmt.Sprintf("/%s/%s/%s", modelV1.PluralKindMap[kind], m.Project, m.Name), nil
+	case *modelV1.Metadata:
+		return fmt.Sprintf("/%s/%s", modelV1.PluralKindMap[kind], m.Name), nil
+	}
+	return "", fmt.Errorf("metadata %T not managed", metadata)
 }
 
-func (d *fileDAO) Create(key string, entity interface{}) error {
+type DAO struct {
+	databaseModel.DAO
+	Folder    string
+	Extension config.FileExtension
+}
+
+func (d *DAO) Create(entity modelAPI.Entity) error {
+	key, generateKeyErr := generateKey(modelV1.Kind(entity.GetKind()), entity.GetMetadata())
+	if generateKeyErr != nil {
+		return generateKeyErr
+	}
 	filePath := d.buildPath(key)
 	if _, err := os.Stat(filePath); err == nil {
 		// The file exists, so we should return a conflict error.
-		// Let's use the etcd error so the caller doesn't have to handle multiple different kind of error
-		// It's an easy hack let's say, but a bit crappy. We should probably at some point defined a higher error to wrap the one coming from the package etcd.
-		return &etcd.Error{Key: key, Code: etcd.ErrorCodeKeyConflict}
+		return &databaseModel.Error{Key: key, Code: databaseModel.ErrorCodeConflict}
 	}
-	return d.Upsert(key, entity)
+	return d.Upsert(entity)
 }
-func (d *fileDAO) Upsert(key string, entity interface{}) error {
+func (d *DAO) Upsert(entity modelAPI.Entity) error {
+	key, generateKeyErr := generateKey(modelV1.Kind(entity.GetKind()), entity.GetMetadata())
+	if generateKeyErr != nil {
+		return generateKeyErr
+	}
 	filePath := d.buildPath(key)
 	if err := os.MkdirAll(filepath.Dir(filePath), 0700); err != nil {
 		return err
@@ -42,18 +73,22 @@ func (d *fileDAO) Upsert(key string, entity interface{}) error {
 	}
 	return os.WriteFile(filePath, data, 0600)
 }
-func (d *fileDAO) Get(key string, entity interface{}) error {
+func (d *DAO) Get(kind modelV1.Kind, metadata modelAPI.Metadata, entity modelAPI.Entity) error {
+	key, generateKeyErr := generateKey(kind, metadata)
+	if generateKeyErr != nil {
+		return generateKeyErr
+	}
 	filePath := d.buildPath(key)
 	data, err := os.ReadFile(filePath)
 	if err != nil {
 		if os.IsNotExist(err) {
-			return &etcd.Error{Key: key, Code: etcd.ErrorCodeKeyNotFound}
+			return &databaseModel.Error{Key: key, Code: databaseModel.ErrorCodeNotFound}
 		}
 		return err
 	}
 	return d.unmarshal(data, entity)
 }
-func (d *fileDAO) Query(query etcd.Query, slice interface{}) error {
+func (d *DAO) Query(query databaseModel.Query, slice interface{}) error {
 	typeParameter := reflect.TypeOf(slice)
 	result := reflect.ValueOf(slice)
 	// to avoid any miss usage when using this method, slice should be a pointer to a slice.
@@ -73,7 +108,7 @@ func (d *fileDAO) Query(query etcd.Query, slice interface{}) error {
 	if typeParameter.Kind() != reflect.Slice {
 		return fmt.Errorf("slice in parameter is not actually a slice but a %q", typeParameter.Kind())
 	}
-	q, err := query.Build()
+	q, err := buildQuery(query)
 	if err != nil {
 		return fmt.Errorf("unable to build the query: %s", err)
 	}
@@ -82,7 +117,7 @@ func (d *fileDAO) Query(query etcd.Query, slice interface{}) error {
 	// Or it's the partial name, and it should only be used to filter the list of the document.
 	// For example: `/projects/per`.
 	// `/projects` is the folder we are looking for, `per` is the partial name.
-	folder := path.Join(d.folder, q)
+	folder := path.Join(d.Folder, q)
 	prefix := ""
 	// An easy way to achieve is to:
 	// 1. try if the path exist with the given query.
@@ -133,41 +168,45 @@ func (d *fileDAO) Query(query etcd.Query, slice interface{}) error {
 	result.Elem().Set(sliceElem)
 	return nil
 }
-func (d *fileDAO) Delete(key string) error {
+func (d *DAO) Delete(kind modelV1.Kind, metadata modelAPI.Metadata) error {
+	key, generateKeyErr := generateKey(kind, metadata)
+	if generateKeyErr != nil {
+		return generateKeyErr
+	}
 	filePath := d.buildPath(key)
 	err := os.Remove(filePath)
 	if err != nil {
 		if os.IsNotExist(err) {
-			return &etcd.Error{Key: key, Code: etcd.ErrorCodeKeyNotFound}
+			return &databaseModel.Error{Key: key, Code: databaseModel.ErrorCodeNotFound}
 		}
 		return err
 	}
 	return nil
 }
 
-func (d *fileDAO) HealthCheck() bool {
+func (d *DAO) HealthCheck() bool {
 	return true
 }
 
-func (d *fileDAO) buildPath(key string) string {
-	return path.Join(d.folder, fmt.Sprintf("%s.%s", key, d.extension))
+func (d *DAO) buildPath(key string) string {
+	return path.Join(d.Folder, fmt.Sprintf("%s.%s", key, d.Extension))
 }
 
-func (d *fileDAO) unmarshal(data []byte, entity interface{}) error {
-	if d.extension == config.JSONExtension {
+func (d *DAO) unmarshal(data []byte, entity interface{}) error {
+	if d.Extension == config.JSONExtension {
 		return json.Unmarshal(data, entity)
 	}
 	return yaml.Unmarshal(data, entity)
 }
 
-func (d *fileDAO) marshal(entity interface{}) ([]byte, error) {
-	if d.extension == config.JSONExtension {
+func (d *DAO) marshal(entity interface{}) ([]byte, error) {
+	if d.Extension == config.JSONExtension {
 		return json.Marshal(entity)
 	}
 	return yaml.Marshal(entity)
 }
 
-func (d *fileDAO) visit(rootPath string, prefix string) ([]string, error) {
+func (d *DAO) visit(rootPath string, prefix string) ([]string, error) {
 	var result []string
 	err := filepath.Walk(rootPath, func(path string, info fs.FileInfo, err error) error {
 		if err != nil {
@@ -177,7 +216,7 @@ func (d *fileDAO) visit(rootPath string, prefix string) ([]string, error) {
 			return nil
 		}
 		fileName := info.Name()
-		if filepath.Ext(fileName) != fmt.Sprintf(".%s", d.extension) {
+		if filepath.Ext(fileName) != fmt.Sprintf(".%s", d.Extension) {
 			// skip every file that doesn't have the correct extension
 			return nil
 		}
